[Ide] Split out FileWatcherService into logical components
